Add(.auth): separate dto from user
