+++ src/main/java/com/newbarams/ajaja/module/user/dto/UserResponse.java
@@ -6,13 +6,6 @@ import lombok.Data;
 
 public final class UserResponse {
 	@Data
-	public static class Token {
-		private final String accessToken;
-		private final String refreshToken;
-		private final long accessTokenExpireIn;
-	}
-
-	@Data
 	public static class MyPage {
 		private final String nickname;
 		private final String defaultEmail;
@@ -22,7 +15,7 @@ public final class UserResponse {
 
 		@QueryProjection
 		public MyPage(String nickname, String defaultEmail, String remindEmail, boolean isEmailVerified,
-			String receiveType) {
+				String receiveType) {
 			this.nickname = nickname;
 			this.defaultEmail = defaultEmail;
 			this.remindEmail = remindEmail;
